+++ src/components/common/KakaoLoginBtn.css
@@ -30,5 +30,5 @@
 }
 
 .btn-kakao:hover {
-  background-color: #fee500a7;
+  background-color: #fff7ac;
 }
