--- src/components/template/AuthRedirect.jsx
@@ -1,10 +1,0 @@
-const AuthRedirect = () => {
-  return (
-    <>
-      로그인 성공창입니다. 리다이렉트 url 설정할때 template AuthRedirect.jsx로
-      해주세요~
-    </>
-  );
-};
-
-export default AuthRedirect;
