+++ src/pages/AuthRedirectPage.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import AuthRedirectTemplate from "../components/template/AuthRedirectTemplate";
+
+const AuthRedirectPage = () => {
+  return <AuthRedirectTemplate />;
+};
+
+export default AuthRedirectPage;
