+++ src/utils/URL.jsx
@@ -1,1 +1,1 @@
-export const BASE_URL = "http://localhost:8080";
+export const BASE_URL = "http://happynewhere.xn--hk3b17f.xn--3e0b707e:8080";
